Warn when applying the RecordCallers attribute to an anonymous sub
or to a lexical sub (also anonymous from the point of view of
symbol tables). This requires some nasty Carp trickery to make
the warning come from the right place (or else it's seen as
coming from attributes.pm). To clean that up it would be nice
to port the module to pure attributes.pm instead of using
Attribute::Handlers.